requirements.txt
pypdf
sentence-transformers
faiss-cpu
rank-bm25
numpy
